Template MIPS function to implement

+++ src/Architecture/Interconnect.h
@@ -12,11 +12,11 @@ namespace ePugStation
 		Interconnect() = default;
 		~Interconnect() = default;
 
-		uint32_t load8(uint32_t address) const;
-		uint32_t load16(uint32_t address) const;
+		uint8_t load8(uint32_t address) const;
+		uint16_t load16(uint32_t address) const;
 		uint32_t load32(uint32_t address) const;
-		void store8(uint32_t address, uint32_t value);
-		void store16(uint32_t address, uint32_t value);
+		void store8(uint32_t address, uint8_t value);
+		void store16(uint32_t address, uint16_t value);
 		void store32(uint32_t address, uint32_t value);
 
 	private:
